+++ azplugins/AnisoPairPotentials.cu
@@ -1,0 +1,23 @@
+// Copyright (c) 2016-2017, Panagiotopoulos Group, Princeton University
+// This file is part of the azplugins project, released under the Modified BSD License.
+
+// Maintainer: mphoward / Everyone is free to add additional potentials
+
+/*!
+ * \file AnisoPairPotentials.cu
+ * \brief Defines the driver functions for computing pair forces on the GPU
+ *
+ * Each pair potential evaluator needs to have an explicit instantiation of the
+ * compute_aniso_pair_potential.
+ */
+
+#include "AnisoPairPotentials.cuh"
+
+namespace azplugins
+{
+namespace gpu
+{
+
+
+} // end namespace gpu
+} // end namespace azplugins