+++ azplugins/module.cc
@@ -11,6 +11,7 @@
 #include "hoomd/extern/pybind/include/pybind11/pybind11.h"
 namespace py = pybind11;
 
+#include "AnisoPairPotentials.h"
 #include "PairPotentials.h"
 #include "WallPotentials.h"
 
@@ -102,6 +103,8 @@ PYBIND11_PLUGIN(_azplugins)
     azplugins::detail::export_pair_potential<azplugins::detail::PairEvaluatorLJ124>(m, "PairPotentialLJ124");
     azplugins::detail::export_pair_potential<azplugins::detail::PairEvaluatorLJ96>(m,"PairPotentialLJ96");
     azplugins::detail::export_pair_potential<azplugins::detail::PairEvaluatorShiftedLJ>(m, "PairPotentialShiftedLJ");
+
+    /* Anisotropic pair potentials */
 
     /* Updaters */
     azplugins::detail::export_TypeUpdater(m);